% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{p}
\alias{p}
\alias{p0}
\alias{pc}
\alias{pcs}
\alias{pcnl}
\alias{pcc}
\alias{pcsc}
\alias{pc_and}
\title{paste variants}
\usage{
p(...)

p0(..., sep = "")

pc(..., sep = "")

pcs(..., sep = "")

pcnl(..., sep = "")

pcc(..., sep = "")

pcsc(..., sep = "")

pc_and(..., sep = "")
}
\arguments{
\item{...}{passed on to \code{paste}}

\item{sep}{passed on to \code{base::paste}}
}
\description{
There are all wrappers around \code{base::paste} with different defaults:
\tabular{llcc}{
   \code{}         \tab \strong{mnemonic}        \tab \strong{\code{collapse=}} \tab \strong{\code{sep=}} \cr
   \code{p()}      \tab paste                    \tab \code{NULL}               \tab \code{" "}   \cr
   \code{p0()}     \tab paste0                   \tab \code{NULL}               \tab \code{""}    \cr
   \code{pc()}     \tab paste collapse           \tab \code{""}                 \tab \code{""}    \cr
   \code{pcs()}    \tab paste collapse space     \tab \code{" "}                \tab \code{""}    \cr
   \code{pcc()}    \tab paste collapse comma     \tab \code{", "}               \tab \code{""}    \cr
   \code{pcsc()}   \tab paste collapse semicolon \tab \code{"; "}               \tab \code{""}    \cr
   \code{pcnl()}   \tab paste collapse newline   \tab \code{"\n"}               \tab \code{""}    \cr
   \code{pc_and()} \tab paste collapse and       \tab \emph{varies}             \tab \code{""}    \cr
}
}
\examples{
x <- head(letters, 3)
y <- tail(letters, 3)
p(x, y)
p0(x, y)
pc(x)
pcs(x)
pcc(x)
pcsc(x)
pcnl(x)
pc_and(x[1])
pc_and(x[1:2])
pc_and(x[1:3])
pc_and(x, y)
}
\seealso{
\code{\link{wrap}} \code{\link{sentence}}
}
