% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{sentence}
\alias{sentence}
\title{sentence}
\usage{
sentence(...)
}
\arguments{
\item{...}{passed on to \code{paste}}
}
\description{
A wrapper around \code{paste} that does some simple string cleaning before returning the result. It
\enumerate{
   \item calls \code{trimws}
   \item make sure each string ends with a single \code{"."}
   \item runs of multiple spaces \code{" "} replaced with one space
   \item runs of multiple periods \code{"."} replaced with one period
   \item runs of multiple commas \code{","} replaced with one comma
   \item spaces before commas or periods removed
}
}
\examples{
sentence(
 "The", c("first", "second", "third"),
   "letter is", letters[1:3], ".",
 "That's important to know")
 x <- p0( "a sentence with   excessive or missing whitespace,",
", extra punctuation, and missing capilitization.more than one in fact ! .Three,actually")
cat(x, "\\n")
cat(sentence(x), "\\n")
}
