% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{wrap}
\alias{wrap}
\alias{dbl_quote}
\alias{sngl_quote}
\alias{bracket}
\alias{parens}
\alias{unparens}
\title{Wrap strings}
\usage{
wrap(x, left = "", right = left)

dbl_quote(x)

sngl_quote(x)

bracket(x)

parens(x)

unparens(x)
}
\arguments{
\item{x}{character to wrap}

\item{left}{character}

\item{right}{character}
}
\description{
These are a series of helpers to help wrap strings with two flanking characters
}
\examples{
wrap("abc", "__")
sngl_quote("abc")
dbl_quote("abc")
parens("abc")
bracket("abc")
(x <- p("name", parens("attribute")))
unparens(x)
}
\seealso{
paste-varients sentence
}
