% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{sentence}
\alias{sentence}
\title{sentence}
\usage{
sentence(...)
}
\arguments{
\item{...}{passed on to \code{paste}}
}
\description{
A wrapper around \code{paste} that does some simple cleaning appropriate for
prose sentences before returning the result. It
\enumerate{
   \item trims leading and trailing whitespace
   \item collapses runs of multiple whitespace into a single space
   \item appends a period \code{.} if there is no sentence ending
         punctuation (\code{?}, \code{!}, or \code{.})
   \item removes spaces preceding punctuation characters: \code{.?!,;}
   \item collapses sequences of punctuation characters (\code{.?!;,}) (possibly
         separated by spaces), into a single punctuation character. The first
         punctuation character of the sequence is used, with priority given to
         sentence ending punctuation \code{.?!}, if present.
   \item makes sure a space follows every \code{.} or \code{,}, (unless
         followed by a digit or the end of the string)
   \item makes sure a space follows every \code{?}, \code{!} or \code{;}
         (unless it's the end of the string)
   \item capitalizes the first letter of each sentence (start of string or
         following a \code{.?!})
}
}
\examples{
sentence(
 "The", c("first", "second", "third"),
   "letter is", letters[1:3], ".",
 "That's important to know")
 x <- p0( "a sentence with   excessive or missing whitespace,",
", extra punctuation, and missing capilitization.more than one in fact ! .Three,actually")
cat(x, "\\n")
cat(sentence(x), "\\n")
}
