% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{wrap}
\alias{wrap}
\alias{dbl_quote}
\alias{sngl_quote}
\alias{bracket}
\alias{brace}
\alias{parens}
\title{Wrap strings}
\usage{
wrap(x, left = "", right = left)

dbl_quote(...)

sngl_quote(...)

bracket(...)

brace(...)

parens(...)
}
\arguments{
\item{x}{character to wrap}

\item{left}{character}

\item{right}{character}

\item{...}{passed to \code{paste} before wrapping}
}
\description{
Wrap strings with flanking characters
}
\examples{
wrap("abc", "__")  #  __abc__
parens("abc")      #   (abc)
sngl_quote("abc")  #   'abc'
dbl_quote("abc")   #   "abc"
bracket("abc")     #   [abc]
brace("abc")       #   {abc}

label <- p("name", parens("attribute"))

label             # "name (attribute)"
unparens(label)   # "name attribute"
}
\seealso{
\code{\link{unwrap}} \code{\link{p}} \code{\link{sentence}}
\code{\link[base]{paste}}
}
